feat: eject emote
funny amongus reference
